Move json pickling into separate JsonSupport trait
.. and fix pickling of FileTree :).
